FIX: Rename pages and components using lowercase.
There were problems on deployment so every .vue file will be named lowercase except components.
