--- src/pages/another/Index.vue
@@ -1,3 +1,0 @@
-<template>
-    <h1>Another</h1>
-</template>
