+++ src/shared/components/Navbar.vue
