+++ src/shared/layouts/default.vue
@@ -6,5 +6,5 @@
 </template>
 
 <script setup lang="ts">
-import Navbar from '../components/navbar.vue';
+import Navbar from '../components/Navbar.vue';
 </script>
